feat: Create new shared resources library
The watch face will use the same colour palette as the main app so colors.xml should be shared between modules. However, the previous library was a Java library and it didn't include resource files. It was replaced with Android Library which holds colour resources as well as Constants class from the previous library.