feat: Improve appearance of short text complication
The complication displays short text in the upper half of the background circle, while the title is placed in the bottom one. The title is drawn with lighter color to put more emphasis on the text itself. Both halves are separated by a line centered between the text. If there is no title, line is not displayed and short text takes centre position in the circle.

 All text in this complication is scaled dynamically depending on the length of the text.